# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Transport Service Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/accomodationPhotos": {
            "post": {
                "description": "Creates a photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodationPhotos"
                ],
                "summary": "Creates a photo",
                "parameters": [
                    {
                        "description": "Post payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateAccomodationPhotoPayload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/store.AccomodationPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodationPhotos/accomodation/{id}": {
            "get": {
                "description": "Fetches photo by a trip id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodationPhotos"
                ],
                "summary": "Fetches photo by a trip id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Accomodation id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AccomodationPhoto"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodationPhotos/accomodationId/{id}": {
            "delete": {
                "description": "Deletes photos by a accomodation id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodationPhotos"
                ],
                "summary": "Deletes photos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "accomodation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodationPhotos/id/{id}": {
            "get": {
                "description": "Fetches a photo by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodationPhotos"
                ],
                "summary": "Fetches a photo by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "photo id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AccomodationPhoto"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Deletes a photo by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodationPhotos"
                ],
                "summary": "Deletes a photo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "photo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodations": {
            "post": {
                "description": "Creates a accomodation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodations"
                ],
                "summary": "Creates a accomodation",
                "parameters": [
                    {
                        "description": "Post payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateAccomodationPayload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/store.Accomodation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodations/id/{id}": {
            "get": {
                "description": "Fetches a accomodation by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodations"
                ],
                "summary": "Fetches a accomodation by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Accomodation id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Accomodation"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/accomodations/tripId/{id}": {
            "get": {
                "description": "Fetches accomodation by a trip id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accomodations"
                ],
                "summary": "Fetches accomodation by a trip id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Accomodation"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.CreateAccomodationPayload": {
            "type": "object",
            "required": [
                "description",
                "name",
                "price_per_night",
                "trip_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_night": {
                    "type": "number"
                },
                "trip_id": {
                    "type": "integer"
                }
            }
        },
        "main.CreateAccomodationPhotoPayload": {
            "type": "object",
            "required": [
                "accomodation_id",
                "photo_url"
            ],
            "properties": {
                "accomodation_id": {
                    "type": "integer"
                },
                "photo_url": {
                    "type": "string"
                }
            }
        },
...
        "store.Accomodation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price_per_night": {
                    "type": "number"
                },
                "trip_id": {
                    "type": "integer"
                }
            }
        },
        "store.AccomodationPhoto": {
            "type": "object",
            "properties": {
                "accomodation_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photo_url": {
                    "type": "string"
                }
            }
        },
...
    }
}